【修改描述】uint64强转uint16,高位截断，ptId定义为uint16_t 【DTS/CR/AR单号】DTS2024101540772 【合入版本】BeiMing 24.2.0.B001,
Match-id-09afa2d09f00af251ef0a1c6367e3db6049f01db

+++ src/cache/read/rcache_flow.h
@@ -19,7 +19,7 @@ public:
 
     ~RCacheFlow();
 
-    BResult Initialize(uint64_t ptId, uint16_t diskId, FlowType flowType, std::vector<uint64_t> flowIds);
+    BResult Initialize(uint16_t ptId, uint16_t diskId, FlowType flowType, std::vector<uint64_t> flowIds);
 
     void Destroy();
 
@@ -55,7 +55,7 @@ public:
 
     DEFINE_REF_COUNT_FUNCTIONS
 private:
-    uint64_t mPtId = 0;
+    uint16_t mPtId = 0;
     uint16_t mDiskId = 0;
 
     FlowPtr mMetaFlow = nullptr;
